Commit for Size is not grandeur

+++ 0x00-hello_world/6-size.c
@@ -7,12 +7,12 @@
  */
 int main(void)
 {
-	printf("size of a char: %lu byte(s)\n",(unsigned long) sizeof(char));
-	printf("size of a int: %lu byte(s)\n",(unsigned long) sizeof(int));
-	printf("size of a long int: %lu byte(s)\n",(unsigned long) sizeof(long int));
-	printf("size of a long long int: %lu byte(s)\n",(unsigned long) sizeof(long long int));
-	printf("size of a long long int: %lu byte(s)\n",(unsigned long) sizeof(long long int));
-	printf("size of a long long int: %lu byte(s)\n",(unsigned long) sizeof(long long int));
-	printf("size of a float: %lu byte(s)\n",(unsigned long) sizeof(float));
+	printf("size of a char: %li byte(s)\n", sizeof(char));
+	printf("size of a int: %li byte(s)\n", sizeof(int));
+	printf("size of a long int: %li byte(s)\n", sizeof(long int));
+	printf("size of a long long int: %li byte(s)\n", sizeof(long long int));
+	printf("size of a long long int: %li byte(s)\n", sizeof(long long int));
+	printf("size of a long long int: %li byte(s)\n", sizeof(long long int));
+	printf("size of a float: %li byte(s)\n", sizeof(float));
 	return (0);
 }
